test(cli): isolate Codex hook alias registry coverage
Move the alias assertion out of the mocked context command suite so
pre-push can load the live registry snapshot.
